Mark keyboard key names translatable in a specific context
Many key names are short and commonly used words; e.g. "Delete", "Left",
"Right", "Pause", "Menu", "End".

It is not at all unlikely that such words could be used in a completely
different meaning elsewhere in the code. That is why this list of
keys should be assigned a context.

The context fulfills also the same purpose as TRANSLATORS: comments --
removing comments as redundant.